backend/requirements.txt
# Backend Dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
pydantic-settings==2.1.0

# Deep Learning
torch==2.1.0
torchvision==0.16.0
timm==0.9.12
efficientnet-pytorch==0.7.1

# Image/Video Processing
opencv-python==4.8.1.78
Pillow==10.1.0
albumentations==1.3.1
face-recognition==1.3.0
dlib==19.24.2

# Explainability
grad-cam==1.4.8
captum==0.6.0

# Utilities
numpy==1.24.3
pandas==2.1.3
scikit-learn==1.3.2
matplotlib==3.8.2
seaborn==0.13.0
tqdm==4.66.1
python-dotenv==1.0.0
aiofiles==23.2.1

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
